requirements.txt
websocket-client>=1.6.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
flask>=3.0.0
flask-socketio>=5.3.0
yfinance>=0.2.36
requests>=2.31.0
